Code_Agent_CLI/claude-tool-agent-example/requirements.txt
anthropic>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
